Revert "move part syncer files"
This reverts commit ca3fab048ebc9e459839bf364559e9b1b52f289b.
