+++ DFC.ServiceTaxonomy.GraphSync/GraphSyncers/GraphSyncPartUriIdProperty.cs
@@ -1,6 +1,6 @@
 using DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces;
 
-namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Parts
+namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers
 {
     public class GraphSyncPartUriIdProperty : IGraphSyncPartIdProperty
     {
